Fixed the issue of javascript being allergic to ;
This may or may not be acceptable solution.

I saved the config coming from the webpage with | instead of ;
and used a python script to copy the data to the "production" config and
then replace | with ;. When I fetch the data, I do the reverse since if
I left the ; in place then, javascript will ignore everything after the first ;.
Since It takes a half second longer to load the server settings. I added a function
to check to see if the page is loading and if it is, it will display a loading page.
I will need to figure out a cool animation to display. I see in hindsight that a javascript
based backend comes with some limitations. That being said, I am sure I can figure out
a way to bring performance back into program. Probably though caching the server settings
on initial load and storing it in localstorage. I will have to see if react has a function
to detect page refreshes. If it does, I can use that to call an api route and fetch the current
server settings and cache them. With that flow, the initial load will be the longest and performance
will come back to the rest of the program.