Another road block but I think I have a solution.
So when editing a mods of a pz server. Some workshop ID have multiple mod names
The issue comes after we save it and want to call it back to the front end.
there is not a real way that I know of that we can persist the relationship. For now,
I plan on saving the data to local storage and then sending it to the server.
I think I will deal with the issue in the backend. It isn't doing much so at least
it will have some work such as converting the data format into a usable form and then
rebuilding it so that the front end can use it. Python will handle it for us.