Have a prototype for the mods page.
I ended up using localstorage to store values and when the mods were
were done being added, I converted it to the form I wanted and sent it on its way.
With this set up, the flow of communication is one way which is not ideal, but
it works and works reasonably well. In the future I would like to have the groupings
saved server side but before then I want to see how big the backend gets and how
much resources it will need to run. Since, it will be located on the server machine,
I want it to be light weight and performant. Time will tell if that is the case.